static 関数置き場 MeshUtility を分配。
