--- Assets/UniGLTF/Editor/MeshUtility/MeshUtility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c1bbe850b95e44740bbbb44064e17d25
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
